Update README.md
This is the project readme for the development and deployment of blackswitch.online. All material, unless specifically noted, is published under the GPL 3.0 and as long as it is followed, you may do anything with this that you wish. This repo is more so that I can follow what I have done and the changes that I have made for this site. I, and I suspect that no one, can remember all that is done to a website as it goes on, and who would want to. If you decide you would like to help, contact me and lets do so. 

Thank you World, nice to see you again. 

Sincerely, 

King Blackswitch
Post Traumatic Studios